requirements.txt
fastapi
pydantic
uvicorn
black
ruff
python-dotenv
httpx
transformers
torch
python-jose
rich
sqlalchemy[asyncio]
asyncpg
cuid